Major project structure cleanup and refactoring
