+++ frontend/frontend/screens/index.tsx
@@ -1,4 +1,4 @@
-export { default as DataStore } from './DataStore/DataStore';
+export { default as DataStore } from '../screens/DataStore/DataStore';
 export { default as ImageStore } from './ImageStore/ImageStore';
 export { default as DocumentStore } from './DocumentStore/DocumentStore';
 export { default as DatabaseAdmin } from './DatabaseAdmin/DatabaseAdmin';
